Move program writing tests into tests crate
The program writing tests were originally located inside of
`chalk-integration`. This moves them to the outer test crate to bring
them in line with existing test practices in chalk.
