+++ Cargo.toml
@@ -32,5 +32,7 @@ chalk-integration = { version = "0.16.0-dev.0", path = "chalk-integration" }
 [workspace]
 
 [dev-dependencies]
+# used for program_writer test errors
+diff = "0.1"
 pretty_assertions = "0.6.1"
 tracing = "0.1"
