+++ tests/lib.rs
@@ -3,6 +3,7 @@ mod test_util;
 #[macro_use]
 mod test;
 
+mod logging_db;
 mod lowering;
 
 mod integration;
