+++ tests/logging_db/lifetimes.rs
@@ -1,0 +1,27 @@
+use super::*;
+#[test]
+fn test_various_forall() {
+    reparse_test(
+        "
+            struct Foo<'b> where forall<'a> Foo<'a>: Baz<'a> { }
+            trait Baz<'a> {}
+            trait Bax<'a> {}
+            trait Biz {
+                type Bex: forall<'a> Bax<'a>;
+            }
+            impl<'a> Baz<'a> for for<'b> fn(Foo<'b>) { }
+            impl<'a> Bax<'a> for fn(Foo<'a>) { }
+            impl<'a> Bax<'a> for dyn forall<'b> Baz<'b> { }
+            ",
+    );
+}
+#[test]
+fn test_lifetimes_in_structs() {
+    reparse_test(
+        "
+            struct Foo<'b> { }
+            trait Baz<'a> {}
+            impl<'a> Baz<'a> for Foo<'a> { }
+            ",
+    );
+}
